requirements.txt
asgiref==3.2.4
Django==3.0.4
django-cors-headers==3.2.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djangorestframework-stubs==1.1.0
gunicorn==20.0.4
numpy==1.18.2
pandas==1.0.3
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
